First pass issues fixed. Correct number of vertices detected
Oh my goodness. It's finally working. The issue was when reshaping the texture buffer after the pass. I was using the default column ordering, but it seems that Fortran/"F" order is required for the format of the moderngl textures. Probably was in the documentation... (Can't find it, but oh well)