Output failed successfully!
Well, it's something. It currently outputs monsterous things, and the larger the resolution, the larger the monster. I need to debug the simplified vertices and see how big they are, cause this may be part of the issue. I believe there is also some discrepancy when indexing into the simplified vertex array, as well. (May have to do with the order of memory in the numpy arrays vs moderngl.buffer...)