Add more printouts to debug_dump(); "Fixed" memory issue with decimated model...
In the process of trying to fix the thirdpass output seemingly consistently degenerating to the error points of vec3(-1), especially more so at lower resolutions levels, more print statements were added to debug_dump.
For some reason, and I do *not* know why it suddenly started working, the vertices -> vec3(-1) stopped. However, when I removed the "large" print statements from between the compute shaders .run() 's, the error came back. I am now assuming the error is in the concurrency of the execution of the individual shaders (i.e. that one can start before the previous one is completely done. This would explain why the vertices to "go" first in the rendered output were the ones furthest from the origin and would be in the furthest cells???? maybe??? I am honestly unsure. But it works for now...