Add switch from avg_vertex only and avg-on-fail in 2nd pass
Current issue: After multiple switches (using "A"), at some point there comes a time when some points (seem to be the same everytime) start to show the average as begin wayyyy to the left. Then some points in the hand go wayyy down. I did not continue to see this pattern, but do know that switching resolutions seems to fix it.