Add face correction weighting to first pass quadric generation
In Hoppe's 1999 paper, he mentions the the Garland and Heckbert error quadric metric and gives its structure, but adds in a face weighting that he says comes from their paper. I can find no mention of this in their paper, but I have added it to my implementation. It doesn't seem to affect the output too much as the triangles are all of a similar size when joined between clusters. A difference can be seen in some of the vertices, but you have to look for it when switching quickly. Maybe I'm just not as sensitive to this effect.
Also added some comments, and made the mesh not scaled down when rendering during animating.
Also also added a final ctx.finish() after the third compute shader, since the next parts mayyyyy start before the data has actually been completed....