Update secondpass.comp to allow for negative determinants
I realized I was not allowing quadrics with negative determinants to pass the "minimum determinant value" test. So I made the test use the absolute value. I do not know if it impacted the output or not, and I am unsure of the implications of a quadric having a negative determinant in the first place.