Change to execute first pass over faces
Last version used the first pass to execute over the clusters, and looped over the faces to see which triangles each cluster contained. This version instead actually uses the atomic addition to execute individually for each face, and each vertex's information is added to the corresponding cluster in the quadric map. (Not sure about the actual impact; I don't remember the original timing results.)